requirements.txt
numpy==1.23.2
Pillow==9.4.0
skimage==0.0
streamlit==1.18.1
